created separate components for each buyer profile accordion
